fix(mail_download):修复无法更新问题

+++ service/system_ser/system.go
@@ -2,13 +2,15 @@ package system_ser
 
 import (
 	"fmt"
+	"github.com/go-git/go-git/v5"
+	"github.com/go-git/go-git/v5/plumbing"
+	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/pkg/errors"
 	"mail_download/model/request_model"
 	"mail_download/model/response_model"
 	"mail_download/tools"
 	customErr "mail_download/tools/error"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"sync"
@@ -52,20 +54,20 @@ func CheckUpdate(param request_model.CheckUpdateParam) (response_model.Version, 
 	if param.Type == 1 {
 		return response_model.Version{CurrentVersion: tools.Version}, nil
 	}
-	return checkUpdate()
+	return checkUpdate("")
 }
 
 var (
 	mu sync.Mutex
 )
 
-func checkUpdate() (response_model.Version, error) {
+func checkUpdate(path string) (response_model.Version, error) {
 	var (
 		response          = response_model.Version{}
 		url               = "https://gitee.com/xiaojiabaoo/mail_download.git"
 		branch            = "main"
 		err               error
-		commit            string
+		commit            = &object.Commit{}
 		index, i          int
 		newVersion        float64
 		exist             bool
@@ -73,11 +75,15 @@ func checkUpdate() (response_model.Version, error) {
 	)
 	mu.Lock()
 	defer mu.Unlock()
-	appPath, err = os.Getwd()
-	if err != nil {
-		return response, customErr.New(customErr.GET_APPPATH_ERROR, "")
+	if path == "" {
+		appPath, err = os.Getwd()
+		if err != nil {
+			return response, customErr.New(customErr.GET_APPPATH_ERROR, "")
+		}
+		tempPath = appPath + "\\Temp"
+	} else {
+		tempPath = path
 	}
-	tempPath = appPath + "\\Temp"
 	exist, err = DirExist(tempPath)
 	if err != nil {
 		return response, err
@@ -90,38 +96,57 @@ func checkUpdate() (response_model.Version, error) {
 	if err != nil && !strings.Contains(err.Error(), "file already exists") {
 		return response, errors.Wrap(err, "检测更新-创建文件夹失败")
 	}
-	cmd := exec.Command("git", "clone", "-b", branch, url, tempPath)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err = cmd.Run(); err != nil {
-		return response, errors.Wrap(err, "检测更新-拉取项目失败")
-	}
-	// 获取最新提交的信息，并从中截取版本号
-	commit, err = GetCommit(url, branch, appPath, tempPath)
+	// 克隆仓库到本地路径，使用浅克隆
+	repo, err := git.PlainClone(tempPath, false, &git.CloneOptions{
+		URL:           url,
+		ReferenceName: plumbing.NewBranchReferenceName(branch),
+		SingleBranch:  true,
+		Depth:         1, // 只克隆最新的 commit
+	})
 	if err != nil {
-		return response, err
+		return response, errors.Wrap(err, "检测更新-拉取项目失败")
+		os.Exit(1)
 	}
-	index = strings.Index(commit, "Version:")
+	// 获取仓库的HEAD引用
+	ref, err := repo.Head()
+	if err != nil {
+		return response, errors.Wrap(err, "检测更新-拉取项目失败")
+		os.Exit(1)
+	}
+	// 获取最新的 commit 对象
+	commit, err = repo.CommitObject(ref.Hash())
+	if err != nil {
+		fmt.Println("Error getting commit object:", err)
+		os.Exit(1)
+	}
+	// 输出最新的 commit 信息
+	//fmt.Println("Latest commit information:")
+	//fmt.Println("Hash:", commit.Hash)
+	//fmt.Println("Author:", commit.Author.Name)
+	//fmt.Println("Email:", commit.Author.Email)
+	//fmt.Println("Date:", commit.Author.When)
+	//fmt.Println("Message:", commit.Message)
+	index = strings.Index(commit.Message, "Version:")
 	if index >= 0 {
-		newVersion = tools.StringToFloat64(strings.ReplaceAll(commit[index:], "Version:", ""))
+		newVersion = tools.StringToFloat64(strings.ReplaceAll(commit.Message[index:], "Version:", ""))
 	}
-	i = strings.Index(commit, ")：")
+	i = strings.Index(commit.Message, ")：")
 	if i == -1 {
-		i = strings.Index(commit, "):")
+		i = strings.Index(commit.Message, "):")
 	}
 	if i >= 0 {
-		response.Describe = commit[i+2 : index]
+		response.Describe = commit.Message[i+2 : index]
 	}
 	response.CurrentVersion = tools.Version
 	response.NewVersion = newVersion
-	os.RemoveAll(tempPath)
+	if path == "" {
+		os.RemoveAll(tempPath)
+	}
 	return response, nil
 }
 
 func Update() error {
 	var (
-		url     = "https://gitee.com/xiaojiabaoo/mail_download.git"
-		branch  = "main"
 		err     error
 		path    string
 		appPath string
@@ -134,7 +159,7 @@ func Update() error {
 	}
 	// 获取上一级目录
 	path = filepath.Join(filepath.Dir(appPath), ".")
-	update, err = checkUpdate()
+	update, err = checkUpdate("")
 	if err != nil {
 		return err
 	}
@@ -166,57 +191,12 @@ func Update() error {
 	if err != nil {
 		return errors.Wrap(err, "创建文件夹失败")
 	}
-	cmd := exec.Command("git", "clone", "-b", branch, url, path)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err = cmd.Run(); err != nil {
+	_, err = checkUpdate(path)
+	if err != nil {
 		return errors.Wrap(err, "拉取项目失败")
 	}
 	DeleteFile(path + "\\.git")
 	return nil
-}
-
-func GetCommit(url, branch, appPath, tempPath string) (string, error) {
-	err := os.Chdir(tempPath)
-	if err != nil {
-		return "", err
-	}
-	defer func() error {
-		err = os.Chdir(appPath)
-		if err != nil {
-			return err
-		}
-		return nil
-	}()
-	cmd := exec.Command("git", "ls-remote", "--quiet", "--tags", "--heads", url, branch)
-	output, err := cmd.Output()
-	if err != nil {
-		return "", errors.Wrap(err, "获取远程存储库信息时出错")
-	}
-	lines := strings.Split(string(output), "\n")
-	if len(lines) <= 0 {
-		return "", nil
-	}
-	fields := strings.Fields(lines[0])
-	if len(fields) <= 1 {
-		return "", nil
-	}
-	commitDescription, err := GetCommitDescribe(fields[0])
-	if err != nil {
-		return "", err
-	}
-	return commitDescription, nil
-}
-
-func GetCommitDescribe(hash string) (string, error) {
-	cmd := exec.Command("git", "show", "-s", "--format=%B", hash)
-	cmd.Stdout = nil // 设置为 nil 避免 Stdout 重复设置
-	cmd.Stderr = os.Stderr
-	output, err := cmd.Output()
-	if err != nil {
-		return "", errors.Wrap(err, "获取提交描述时出错")
-	}
-	return strings.TrimSpace(string(output)), nil
 }
 
 func DeleteFile(filePath string) error {
